+++ examples/scigrid-de/scigrid-lopf-then-pf-plotly.py
@@ -22,7 +22,7 @@
 #
 # Wind and solar capacities and locations: EEG Stammdaten, based on  http://www.energymap.info/download.html, which represents capacities at the end of 2014. Units without PLZ are removed.
 #
-# Wind and solar time series: REatlas, Andresen et al, "Validation of Danish wind time series from a new global renewable energy atlas for energy system analysis," Energy 93 (2015) 1074 - 1088.
+# Wind and solar time series: REatlas, Andresen et al, "Validation of Danish wind time series from a new global renewable energy atlas for energy system analysis, " Energy 93 (2015) 1074 - 1088.
 #
 # NB:
 #
@@ -107,7 +107,7 @@ network.generators.groupby("carrier")["p_nom"].sum()
 
 network.storage_units.groupby("carrier")["p_nom"].sum()
 
-tech = "Wind Onshore"  # in ["Gas","Brown Coal","Hard Coal","Wind Offshore","Wind Onshore","Solar"]
+tech = "Wind Onshore"  # in ["Gas", "Brown Coal", "Hard Coal", "Wind Offshore", "Wind Onshore", "Solar"]
 
 gens = network.generators[network.generators.carrier == tech]
 gen_distribution = (
@@ -179,7 +179,7 @@ for i in range(int(24 / group_size)):
     network.lines.s_nom = network.lines.s_nom_opt
 
 # if lines are extended, look at which ones are bigger
-# network.lines[["s_nom_original","s_nom"]][abs(network.lines.s_nom - contingency_factor*network.lines.s_nom_original) > 1]
+# network.lines[["s_nom_original", "s_nom"]][abs(network.lines.s_nom - contingency_factor*network.lines.s_nom_original) > 1]
 
 p_by_carrier = network.generators_t.p.groupby(network.generators.carrier, axis=1).sum()
 
@@ -339,7 +339,7 @@ network.generators_t.p_set = network.generators_t.p
 network.generators.control = "PV"
 
 # set slack
-# network.generators.loc["1 Coal","control"] = "Slack"
+# network.generators.loc["1 Coal", "control"] = "Slack"
 
 
 # Need some PQ buses so that Jacobian doesn't break