+++ test/test_pf_against_pandapower.py
@@ -1,42 +1,24 @@
 # -*- coding: utf-8 -*-
 import numpy as np
 import pandapower as pp
+import pytest
 from numpy.testing import assert_array_almost_equal as equal
 
 import pypsa
 
 
-def test_pandapower_case():
-    # more complicated examples like
-    # net = pandapower.networks.example_simple()
-    # can be used once the import of e.g. switches is perfected
-
-    # create empty net
-    net = pp.create_empty_network()
-
-    # create buses
-    b1 = pp.create_bus(net, vn_kv=20.0, name="Bus 1")
-    b2 = pp.create_bus(net, vn_kv=0.4, name="Bus 2")
-    b3 = pp.create_bus(net, vn_kv=0.4, name="Bus 3")
-
-    # create bus elements
-    pp.create_ext_grid(net, bus=b1, vm_pu=1.02, name="Grid Connection")
-    pp.create_load(net, bus=b3, p_mw=0.1, q_mvar=0.05, name="Load")
-
-    # create branch elements
-    pp.create_transformer(
-        net, hv_bus=b1, lv_bus=b2, std_type="0.4 MVA 20/0.4 kV", name="Trafo"
-    )
-    pp.create_line(
-        net, from_bus=b2, to_bus=b3, length_km=0.1, name="Line", std_type="NAYY 4x50 SE"
-    )
-
+@pytest.mark.parametrize("use_pandapower_index", [True, False])
+@pytest.mark.parametrize("extra_line_data", [True, False])
+def test_pandapower_custom_case(
+    pandapower_custom_network, use_pandapower_index, extra_line_data
+):
+    net = pandapower_custom_network
     # because of phase angles, need to init with DC
     pp.runpp(net, calculate_voltage_angles=True, init="dc")
-
     n = pypsa.Network()
-
-    n.import_from_pandapower_net(net)
+    n.import_from_pandapower_net(
+        net, use_pandapower_index=use_pandapower_index, extra_line_data=extra_line_data
+    )
 
     # seed PF with LPF solution because of phase angle jumps
     n.lpf()